move ToVersionProto func
