+++ internal/api/version.go
@@ -1,13 +1,14 @@
-package svcv1alpha1
+package api
 
 import (
 	"google.golang.org/protobuf/types/known/timestamppb"
 
+	svcv1alpha1 "github.com/akuity/kargo/api/service/v1alpha1"
 	"github.com/akuity/kargo/pkg/x/version"
 )
 
-func ToVersionProto(v version.Version) *VersionInfo {
-	return &VersionInfo{
+func ToVersionProto(v version.Version) *svcv1alpha1.VersionInfo {
+	return &svcv1alpha1.VersionInfo{
 		Version:      v.Version,
 		GitCommit:    v.GitCommit,
 		GitTreeDirty: v.GitTreeDirty,
